+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/review/application/usecase/ReviewUseCaseImpl.java
@@ -1,7 +1,11 @@
 package com.romanticpipe.reviewcanvas.domain.review.application.usecase;
 
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.support.TransactionTemplate;
+
 import com.romanticpipe.reviewcanvas.admin.domain.ShopAdmin;
-import com.romanticpipe.reviewcanvas.admin.service.ShopAdminValidator;
+import com.romanticpipe.reviewcanvas.admin.service.ShopAdminService;
 import com.romanticpipe.reviewcanvas.cafe24.product.Cafe24ProductClient;
 import com.romanticpipe.reviewcanvas.cafe24.product.Cafe24ProductDto;
 import com.romanticpipe.reviewcanvas.domain.Product;
@@ -19,16 +23,14 @@ import com.romanticpipe.reviewcanvas.service.ProductValidator;
 import com.romanticpipe.reviewcanvas.service.ReviewCreator;
 import com.romanticpipe.reviewcanvas.service.ReviewReader;
 import com.romanticpipe.reviewcanvas.service.ReviewValidator;
+
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.support.TransactionTemplate;
 
 @Component
 @RequiredArgsConstructor
 class ReviewUseCaseImpl implements ReviewUseCase {
 
-	private final ShopAdminValidator shopAdminValidator;
+	private final ShopAdminService shopAdminService;
 	private final ProductCreator productCreator;
 	private final ProductValidator productValidator;
 	private final ProductReader productReader;
@@ -40,8 +42,8 @@ class ReviewUseCaseImpl implements ReviewUseCase {
 
 	@Override
 	public PageResponse<GetReviewForUserResponse> getReviewsForUser(String mallId, Long productNo,
-																	PageableRequest pageableRequest,
-																	ReviewFilter filter) {
+		PageableRequest pageableRequest,
+		ReviewFilter filter) {
 		Product product = productReader.findProduct(mallId, productNo)
 			.orElseGet(() -> createProduct(mallId, productNo));
 
@@ -75,19 +77,19 @@ class ReviewUseCaseImpl implements ReviewUseCase {
 	public void createReview(String productId, CreateReviewRequest createReviewRequest) {
 		Product product = productValidator.validByProductId(productId);
 		// TODO: product가  mallId와 productNo로 상품을 생성하는 로직을 추가하도록 변경해야 함.
-//		Product product = productReader.findByMallIdAndProductNo(mallId, productNo)
-//			.orElseGet(() -> createProduct(mallId, productNo));
-		ShopAdmin shopAdmin = shopAdminValidator.validById(product.getShopAdminId());
+		//		Product product = productReader.findByMallIdAndProductNo(mallId, productNo)
+		//			.orElseGet(() -> createProduct(mallId, productNo));
+		ShopAdmin shopAdmin = shopAdminService.validById(product.getShopAdminId());
 		// TODO: 프론트로부터 memberId를 받아 user를 조회하여 userId를 가져온다.
-//		Review review = new Review(
-//			null,
-//			null,
-//			createReviewRequest.content(),
-//			createReviewRequest.score(),
-//			shopAdmin.isApproveStatus()
-//				? ReviewStatus.WAITING : ReviewStatus.APPROVED
-//		);
-//		reviewCreator.save(review);
+		//		Review review = new Review(
+		//			null,
+		//			null,
+		//			createReviewRequest.content(),
+		//			createReviewRequest.score(),
+		//			shopAdmin.isApproveStatus()
+		//				? ReviewStatus.WAITING : ReviewStatus.APPROVED
+		//		);
+		//		reviewCreator.save(review);
 	}
 
 	private Product createProduct(String mallId, Long productNo) {
@@ -95,7 +97,7 @@ class ReviewUseCaseImpl implements ReviewUseCase {
 		cafe24ProductDto.validateIsFullContent(mallId, productNo);
 
 		return writeTransactionTemplate.execute(status -> {
-			ShopAdmin shopAdmin = shopAdminValidator.validByMallId(mallId);
+			ShopAdmin shopAdmin = shopAdminService.validByMallId(mallId);
 			Product product = new Product(productNo, cafe24ProductDto.product().productName(), shopAdmin.getId());
 			return productCreator.save(product);
 		});
