--- domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/admin/service/ShopAdminCreator.java
@@ -1,16 +1,0 @@
-package com.romanticpipe.reviewcanvas.admin.service;
-
-import com.romanticpipe.reviewcanvas.admin.domain.ShopAdmin;
-import com.romanticpipe.reviewcanvas.admin.repository.ShopAdminRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class ShopAdminCreator {
-	private final ShopAdminRepository shopAdminRepository;
-
-	public ShopAdmin save(ShopAdmin shopAdmin) {
-		return shopAdminRepository.save(shopAdmin);
-	}
-}
