+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/admin/service/ShopAdminService.java
@@ -1,17 +1,35 @@
 package com.romanticpipe.reviewcanvas.admin.service;
 
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.stereotype.Service;
+
 import com.romanticpipe.reviewcanvas.admin.domain.ShopAdmin;
 import com.romanticpipe.reviewcanvas.admin.exception.AdminNotFoundException;
 import com.romanticpipe.reviewcanvas.admin.exception.ShopAdminErrorCode;
 import com.romanticpipe.reviewcanvas.admin.repository.ShopAdminRepository;
 import com.romanticpipe.reviewcanvas.exception.BusinessException;
+
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class ShopAdminValidator {
+public class ShopAdminService {
+
 	private final ShopAdminRepository shopAdminRepository;
+
+	public ShopAdmin save(ShopAdmin shopAdmin) {
+		return shopAdminRepository.save(shopAdmin);
+	}
+
+	public List<ShopAdmin> findAll() {
+		return shopAdminRepository.findAll();
+	}
+
+	public Optional<ShopAdmin> findByMallId(String mallId) {
+		return shopAdminRepository.findByMallId(mallId);
+	}
 
 	public ShopAdmin validByEmail(String email) {
 		return shopAdminRepository.findByEmail(email)
